advent-py/pyproject.toml
[tool.poetry]
name = "advent-of-code"
version = "0.1.0"
description = ""
authors = ["Caroline Rodewig <16093297+carodewig@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
attrs = "^21.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
